Support for AArch64 compilation
Note, this is a long way from working! It compiles, but doesn't even
attempt to link (and many functions are not yet implemented, such as
anything MMU-related). But from a purely compilation point of view, we now
have basic 64-bit AArch64 support.

+++ k/kipc.c
@@ -21,7 +21,7 @@ uintptr ipc_mapNewSharedPageInCurrentProcess() {
 	// Now map it. Since all involved processes get the same address for a given shared
 	// page, we don't need any additional means of referencing the page, and we don't
 	// even need to track its physical address.
-	uint32 userPtr = userAddressForSharedPage(idx);
+	uintptr userPtr = userAddressForSharedPage(idx);
 	Process* owner = TheSuperPage->currentProcess;
 	bool ok = mmu_newSharedPage(Al, owner, userPtr);
 	if (!ok) {
